refactor: resource delete event (DEV-4468)
