+++ libs/vre/resource-editor/resource-properties/src/lib/erase-resource-dialog.component.ts
@@ -3,6 +3,7 @@ import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { DeleteResource, DeleteResourceResponse, KnoraApiConnection, ReadResource } from '@dasch-swiss/dsp-js';
 import { DspApiConnectionToken } from '@dasch-swiss/vre/core/config';
+import { OntologyService } from '@dasch-swiss/vre/shared/app-helper-services';
 import { finalize } from 'rxjs/operators';
 
 @Component({
@@ -52,6 +53,7 @@ export class EraseResourceDialogComponent {
     private _dspApiConnection: KnoraApiConnection,
     @Inject(MAT_DIALOG_DATA)
     public data: ReadResource,
+    private _ontologyService: OntologyService,
     private _dialogRef: MatDialogRef<EraseResourceDialogComponent>
   ) {}
 
@@ -73,6 +75,7 @@ export class EraseResourceDialogComponent {
         })
       )
       .subscribe(response => {
+        this._ontologyService.updateClassItemCount(this.data);
         this._dialogRef.close(response as DeleteResourceResponse);
       });
   }
