+++ libs/vre/shared/app-helper-services/src/lib/ontology.service.ts
@@ -4,9 +4,12 @@ import {
   Constants,
   KnoraApiConfig,
   ReadOntology,
+  ReadResource,
   ResourcePropertyDefinitionWithAllLanguages,
 } from '@dasch-swiss/dsp-js';
 import { DspApiConfigToken, RouteConstants } from '@dasch-swiss/vre/core/config';
+import { LoadClassItemsCountAction, ProjectsSelectors } from '@dasch-swiss/vre/core/state';
+import { Store } from '@ngxs/store';
 import { Observable, of } from 'rxjs';
 import { DefaultProperties, DefaultProperty, PropertyCategory } from './default-data/default-properties';
 
@@ -22,7 +25,8 @@ export class OntologyService {
 
   constructor(
     @Inject(DspApiConfigToken) private _dspApiConfig: KnoraApiConfig,
-    private _route: ActivatedRoute
+    private _route: ActivatedRoute,
+    private _store: Store
   ) {}
 
   /**
@@ -153,5 +157,16 @@ export class OntologyService {
     }
 
     return ontologyName ? `${iriBase}/${RouteConstants.ontology}/${projectShortcode}/${ontologyName}/v2` : null;
+  }
+
+  updateClassItemCount(resource: ReadResource) {
+    const currentProject = this._store.selectSnapshot(ProjectsSelectors.currentProject);
+    if (currentProject) {
+      const ontologyIri = this.getOntologyIriFromRoute(currentProject.shortcode);
+      if (ontologyIri) {
+        const classId = resource.entityInfo.classes[resource.type]?.id;
+        this._store.dispatch(new LoadClassItemsCountAction(ontologyIri, classId));
+      }
+    }
   }
 }
